+++ src/App.js
@@ -4,7 +4,7 @@ import Banner from './components/Banner';
 import Container from '@mui/material/Container';
 import Home from './components/Home';
 import ComingSoon from './components/ComingSoon';
-import hvacLogo from './assets/hvac-logo.svg';
+import Slider from './components/Slider';
 import './styles/App.css';
 
 function App() {
@@ -12,13 +12,7 @@ function App() {
     <Router>
       <Container maxWidth={false} className="App" sx={{ width: '90vw', minHeight: '90vh', margin: 'auto', flexGrow: 1, display: 'flex', flexDirection: 'column', p: 0, bgcolor: '#fff' }}>
         <Banner />
-        {/* <header className="App-header">
-          <img src={hvacLogo} alt="HVAC Logo" style={{ width: 120, height: 120, marginBottom: 24 }} />
-          <h1>Preferred Services Heating & Air</h1>
-          <p>
-            Your comfort is our priority. We provide professional heating and air services.
-          </p>
-        </header> */}
+        <Slider />
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/service" element={<ComingSoon title="Service" />} />
